outlook-agent/pyproject.toml
[tool.poetry]
name = "outlook-agent"
version = "0.1.0"
description = "LangChain-based Outlook email agent"
authors = ["MKZuziak <maciej.k.zuziak@protonmail.com>"]
readme = "README.md"
packages = [{ include = "outlook-agent", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.13,<4.0"
langchain = "^0.3.27"
openai = "^1.98.0"
exchangelib = "^5.5.1"
python-dotenv = "^1.1.1"

[tool.poetry.group.gui.dependencies]
gradio = "^5.39.0"

[tool.poetry.group.api.dependencies]
fastapi = "^0.116.1"
uvicorn = "^0.35.0"

[tool.poetry.scripts]
email-agent = "outlook-agent.main:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"